+++ lib/commands.ps1
@@ -20,7 +20,7 @@ function command_path($cmd) {
         $shim_path = "$scoopdir\shims\scoop-$cmd.ps1"
         $line = ((Get-Content $shim_path) | Where-Object { $_.startswith('$path') })
         if($line) {
-            Invoke-Expression -command "$line"
+            Invoke-Command ([scriptblock]::Create($line)) -NoNewScope
             $cmd_path = $path
         }
         else { $cmd_path = $shim_path }