+++ lib/core.ps1
@@ -9,9 +9,9 @@ function Optimize-SecurityProtocol {
 
     # If not, change it to support TLS 1.2
     if (!($isNewerNetFramework -and $isSystemDefault)) {
-        # Set to TLS 1.2 (3072), then TLS 1.1 (768), and TLS 1.0 (192). Ssl3 has been superseded,
-        # https://docs.microsoft.com/en-us/dotnet/api/system.net.securityprotocoltype?view=netframework-4.5
-        [System.Net.ServicePointManager]::SecurityProtocol = 3072 -bor 768 -bor 192
+        # Set to TLS 1.2 (3072). Ssl3, TLS 1.0, and 1.1 have been deprecated,
+        # https://datatracker.ietf.org/doc/html/rfc8996
+        [System.Net.ServicePointManager]::SecurityProtocol = 3072
     }
 }
 
@@ -318,6 +318,39 @@ function Get-HelperPath {
     }
 }
 
+function Get-CommandPath {
+    [CmdletBinding()]
+    [OutputType([String])]
+    param(
+        [Parameter(Mandatory = $true, ValueFromPipeline = $true)]
+        [String]
+        $Command
+    )
+
+    begin {
+        $userShims = Convert-Path (shimdir $false)
+        $globalShims = fullpath (shimdir $true) # don't resolve: may not exist
+    }
+
+    process {
+        try {
+            $comm = Get-Command $Command -ErrorAction Stop
+        } catch {
+            return $null
+        }
+        $commandPath = if ($comm.Path -like "$userShims*" -or $comm.Path -like "$globalShims*") {
+            Get-ShimTarget ($comm.Path -replace '\.exe$', '.shim')
+        } elseif ($comm.CommandType -eq 'Application') {
+            $comm.Source
+        } elseif ($comm.CommandType -eq 'Alias') {
+            Get-CommandPath $comm.ResolvedCommandName
+        } else {
+            $null
+        }
+        return $commandPath
+    }
+}
+
 function Test-HelperInstalled {
     [CmdletBinding()]
     param(
@@ -599,6 +632,20 @@ function get_app_name_from_shim($shim) {
     return get_app_name $content
 }
 
+function Get-ShimTarget($ShimPath) {
+    if ($ShimPath) {
+        $shimTarget = if ($ShimPath.EndsWith('.shim')) {
+            (Get-Content -Path $ShimPath | Select-Object -First 1).Replace('path = ', '').Replace('"', '')
+        } else {
+            ((Select-String -Path $ShimPath -Pattern '^(?:@rem|#)\s*(.*)$').Matches.Groups | Select-Object -Index 1).Value
+        }
+        if (!$shimTarget) {
+            $shimTarget = ((Select-String -Path $ShimPath -Pattern '[''"]([^@&]*?)[''"]' -AllMatches).Matches.Groups | Select-Object -Last 1).Value
+        }
+        $shimTarget | Convert-Path
+    }
+}
+
 function warn_on_overwrite($shim, $path) {
     if (!(Test-Path $shim)) {
         return
@@ -726,7 +773,7 @@ function shim($path, $global, $name, $arg) {
     } else {
         warn_on_overwrite "$shim.cmd" $path
         # find path to Git's bash so that batch scripts can run bash scripts
-        $gitdir = (Get-Item (Get-Command git -CommandType:Application -ErrorAction:Stop).Source -ErrorAction:Stop).Directory.Parent
+        $gitdir = (Get-Item (Get-CommandPath git) -ErrorAction:Stop).Directory.Parent
         if ($gitdir.FullName -imatch 'mingw') {
             $gitdir = $gitdir.Parent
         }
@@ -895,7 +942,7 @@ function applist($apps, $global) {
 }
 
 function parse_app([string] $app) {
-    if($app -match '(?:(?<bucket>[a-zA-Z0-9-]+)\/)?(?<app>.*\.json$|[a-zA-Z0-9-_.]+)(?:@(?<version>.*))?') {
+    if($app -match '(?:(?<bucket>[a-zA-Z0-9-_.]+)\/)?(?<app>.*\.json$|[a-zA-Z0-9-_.]+)(?:@(?<version>.*))?') {
         return $matches['app'], $matches['bucket'], $matches['version']
     }
     return $app, $null, $null